requirements.txt
pandas
pytest
streamlit

# RAG dependencies (VibeFinder 2.0)
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
groq>=0.9.0
numpy>=1.24.0
